refactor(events): move domain events to the contexts that raise them
All ten events sat in common/domain/event, named for aggregates they did not
live beside: HealthUpgradeCreated was not in the upgrade context, StreakAchieved
not in tracking. Because the bounded-context rule ignores common wholesale, this
made the slice test pass by relocation rather than by decoupling -- tracking and
reflection could publish upgrade-flavoured events with the guard staying green.
It also made the shared kernel a write-shared surface three contexts edited, and
gave common a dependency on every context via the event handler.

Each event moves to its owning context. Only the DomainEvent marker stays
shared, and DomainEventPublisher moves to common/domain/port/out, where it is
finally covered by the "outbound ports are interfaces" rule -- it was the one
outbound port sitting outside a port package.

Published events are a context's published language, so ..domain.event.. joins
the sanctioned cross-context surfaces. That is a real narrowing overall: consumers
must now name the context they are consuming from, rather than everything
arriving through common.

Deletes EventHandler, ten handlers whose entire body was a debug log, and logs
publication once in SpringDomainEventPublisher instead. That cannot fall behind
the event vocabulary the way a per-type handler list does, records already carry
their payload in toString, and it removes the last common -> context dependency
outside common/security and common/websocket.
