+++ backend/src/main/java/com/healthupgrades/reflection/domain/event/ReflectionAdded.java
@@ -1,4 +1,6 @@
-package com.healthupgrades.common.domain.event;
+package com.healthupgrades.reflection.domain.event;
+
+import com.healthupgrades.common.domain.event.DomainEvent;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
