+++ backend/src/main/java/com/healthupgrades/tracking/application/TrackingService.java
@@ -1,8 +1,8 @@
 package com.healthupgrades.tracking.application;
 
-import com.healthupgrades.common.domain.event.DomainEventPublisher;
-import com.healthupgrades.common.domain.event.ProgressEntryRecorded;
-import com.healthupgrades.common.domain.event.StreakAchieved;
+import com.healthupgrades.common.domain.port.out.DomainEventPublisher;
+import com.healthupgrades.tracking.domain.event.ProgressEntryRecorded;
+import com.healthupgrades.tracking.domain.event.StreakAchieved;
 import com.healthupgrades.common.domain.exception.DuplicateProgressException;
 import com.healthupgrades.common.domain.exception.ResourceNotFoundException;
 import com.healthupgrades.tracking.adapter.in.web.ProgressDto;
