+++ backend/src/main/java/com/healthupgrades/upgrade/domain/event/HealthUpgradeActivated.java
@@ -1,4 +1,6 @@
-package com.healthupgrades.common.domain.event;
+package com.healthupgrades.upgrade.domain.event;
+
+import com.healthupgrades.common.domain.event.DomainEvent;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
